Fix errors in moving module from release-2.x branch
