+++ log4j-jakarta-web/src/main/java/org/apache/logging/log4j/web/WebLookup.java
@@ -21,8 +21,8 @@ package org.apache.logging.log4j.web;
 import jakarta.servlet.ServletContext;
 
 import org.apache.logging.log4j.core.LogEvent;
-import org.apache.logging.log4j.core.config.plugins.Plugin;
 import org.apache.logging.log4j.core.lookup.AbstractLookup;
+import org.apache.logging.log4j.plugins.Plugin;
 import org.apache.logging.log4j.util.Strings;
 
 @Plugin(name = "web", category = "Lookup")
